Modularised fits
The notebook now includes code to fit to all the datasets and work out confidence intervals. Lots of the code has been modularised and the plotting code has been turned into functions to make the notebook more tidy and readable.  The plotting of fitting behaviour still needs to be updated to reflect the new modularised format of the rest of the code.